Fix #38: Added missing documentation about REST API calls - PowerAuthTokenStore.ts moved to lib folder - Fixed various indentation and documentation bugs
