+++ lib/PowerAuth.d.ts
@@ -10,7 +10,7 @@ import { PowerAuthActivation } from './model/PowerAuthActivation';
 import { PowerAuthBiometryInfo } from './model/PowerAuthBiometryInfo';
 import { PowerAuthRecoveryActivationData } from './model/PowerAuthRecoveryActivationData';
 import { PowerAuthConfirmRecoveryCodeDataResult } from './model/PowerAuthConfirmRecoveryCodeDataResult';
-import { PowerAuthTokenStore } from "./core/PowerAuthTokenStore";
+import { PowerAuthTokenStore } from "./PowerAuthTokenStore";
 /**
  * Class used for the main interaction with the PowerAuth SDK components.
  */
@@ -97,11 +97,17 @@ export declare class PowerAuth {
     /**
      * Fetch the activation status for current activation.
      *
+     * The following calls to PowerAuth Standard RESTful API endpoints are performed on the background:
+     * - `/pa/activation/status` - to get the activation status
+     * - `/pa/upgrade/start` - (optional) in case that protocol upgrade is required.
+     * - `/pa/upgrade/commit` - (optional) in case that protocol upgrade is required.
+     * - `/pa/signature/validate` - (optional) as a prevention to local counter desynchronization.
+     *
      * @return A promise with activation status result - it contains status information in case of success and error in case of failure.
      */
     fetchActivationStatus(): Promise<PowerAuthActivationStatus>;
     /**
-     * Create a new activation.
+     * Create a new activation by calling a PowerAuth Standard RESTful API endpoint `/pa/activation/create`.
      *
      * @param activation A PowerAuthActivation object containg all information required for the activation creation.
      */
@@ -121,19 +127,23 @@ export declare class PowerAuth {
      */
     getActivationFingerprint(): Promise<string>;
     /**
-     * Remove current activation by calling a PowerAuth Standard RESTful API endpoint '/pa/activation/remove'.
+     * Remove current activation by calling a PowerAuth Standard RESTful API endpoint `/pa/activation/remove`. The user can authenticate with password
+     * or with biometry, if biometric factor is configured in the current activation. In case of biometry, the system biometric authentication dialog
+     * is displayed, so the operation may take an undefined amount of time to complete.
      *
      * @param authentication An authentication instance specifying what factors should be used to sign the request.
      */
     removeActivationWithAuthentication(authentication: PowerAuthAuthentication): Promise<void>;
     /**
      * This method removes the activation session state and biometry factor key. Cached possession related key remains intact.
-     * Unlike the `removeActivationWithAuthentication`, this method doesn't inform server about activation removal. In this case
+     * Unlike the `removeActivationWithAuthentication()`, this method doesn't inform server about activation removal. In this case
      * user has to remove the activation by using another channel (typically internet banking, or similar web management console)
      */
     removeActivationLocal(): Promise<void>;
     /**
      * Compute the HTTP signature header for GET HTTP method, URI identifier and HTTP query parameters using provided authentication information.
+     * Be aware that if `PowerAuthAuthentication.useBiometry` is true, then the system biometric authentication dialog is displayed, so the operation
+     * may take an undefined amount of time to complete.
      *
      * @param authentication An authentication instance specifying what factors should be used to sign the request.
      * @param uriId URI identifier.
@@ -143,6 +153,8 @@ export declare class PowerAuth {
     requestGetSignature(authentication: PowerAuthAuthentication, uriId: string, params?: any): Promise<PowerAuthAuthorizationHttpHeader>;
     /**
      * Compute the HTTP signature header for given HTTP method, URI identifier and HTTP request body using provided authentication information.
+     * Be aware that if `PowerAuthAuthentication.useBiometry` is true, then the system biometric authentication dialog is displayed, so the operation
+     * may take an undefined amount of time to complete.
      *
      * @param authentication An authentication instance specifying what factors should be used to sign the request.
      * @param method HTTP method used for the signature computation.
@@ -152,7 +164,9 @@ export declare class PowerAuth {
      */
     requestSignature(authentication: PowerAuthAuthentication, method: string, uriId: string, body?: string): Promise<PowerAuthAuthorizationHttpHeader>;
     /**
-     * Compute the offline signature for given HTTP method, URI identifier and HTTP request body using provided authentication information.
+     * Compute the offline signature for given HTTP method, URI identifier and HTTP request body using provided authentication information. Be aware that if
+     * `PowerAuthAuthentication.useBiometry` is true, then the system biometric authentication dialog is displayed, so the operation may take an undefined
+     * amount of time to complete.
      *
      * @param authentication An authentication instance specifying what factors should be used to sign the request. The possession and knowledge is recommended.
      * @param uriId URI identifier.
@@ -166,11 +180,11 @@ export declare class PowerAuth {
      *
      * @param data An arbitrary data
      * @param signature A signature calculated for data, in Base64 format
-     * @param masterKey If true, then master server public key is used for validation, otherwise personalized server's public key.
+     * @param masterKey If `true`, then master server public key is used for validation, otherwise personalized server's public key.
      */
     verifyServerSignedData(data: string, signature: string, masterKey: boolean): Promise<boolean>;
     /**
-     * Change the password, validate old password by calling a PowerAuth Standard RESTful API endpoint '/pa/signature/validate'.
+     * Change the password, validate old password by calling a PowerAuth Standard RESTful API endpoint `/pa/signature/validate`.
      *
      * @param oldPassword Old password, currently set to store the data.
      * @param newPassword New password, to be set in case authentication with old password passes.
@@ -190,7 +204,7 @@ export declare class PowerAuth {
     unsafeChangePassword(oldPassword: string, newPassword: string): Promise<boolean>;
     /**
      * Regenerate a biometry related factor key.
-     * This method calls PowerAuth Standard RESTful API endpoint '/pa/vault/unlock' to obtain the vault encryption key used for original private key decryption.
+     * This method calls PowerAuth Standard RESTful API endpoint `/pa/vault/unlock` to obtain the vault encryption key used for original private key decryption.
      *
      * @param password Password used for authentication during vault unlocking call.
      * @param title (used only in Android) Title for biometry dialog
@@ -215,9 +229,10 @@ export declare class PowerAuth {
      */
     getBiometryInfo(): Promise<PowerAuthBiometryInfo>;
     /**
-     * Generate a derived encryption key with given index.
-     * The key is returned in form of base64 encoded string.
-     * This method calls PowerAuth Standard RESTful API endpoint '/pa/vault/unlock' to obtain the vault encryption key used for subsequent key derivation using given index.
+     * Generate a derived encryption key with given index. The key is returned in form of base64 encoded string.
+     *
+     * This method calls PowerAuth Standard RESTful API endpoint `/pa/vault/unlock` to obtain the vault encryption key used
+     * for subsequent key derivation using given index.
      *
      * @param authentication Authentication used for vault unlocking call.
      * @param index Index of the derived key using KDF.
@@ -225,7 +240,9 @@ export declare class PowerAuth {
     fetchEncryptionKey(authentication: PowerAuthAuthentication, index: number): Promise<string>;
     /**
      * Sign given data with the original device private key (asymetric signature).
-     * This method calls PowerAuth Standard RESTful API endpoint '/pa/vault/unlock' to obtain the vault encryption key used for private key decryption. Data is then signed using ECDSA algorithm with this key and can be validated on the server side.
+     *
+     * This method calls PowerAuth Standard RESTful API endpoint `/pa/vault/unlock` to obtain the vault encryption key
+     * used for private key decryption. Data is then signed using ECDSA algorithm with this key and can be validated on the server side.
      *
      * @param authentication Authentication used for vault unlocking call.
      * @param data Data to be signed with the private key.
@@ -233,7 +250,8 @@ export declare class PowerAuth {
     signDataWithDevicePrivateKey(authentication: PowerAuthAuthentication, data: string): Promise<string>;
     /**
      * Validate a user password.
-     * This method calls PowerAuth Standard RESTful API endpoint '/pa/signature/validate' to validate the signature value.
+     *
+     * This method calls PowerAuth Standard RESTful API endpoint `/pa/signature/validate` to validate the signature value.
      *
      * @param password Password to be verified.
      */
@@ -244,13 +262,15 @@ export declare class PowerAuth {
     hasActivationRecoveryData(): Promise<boolean>;
     /**
      * Get an activation recovery data.
-     * This method calls PowerAuth Standard RESTful API endpoint '/pa/vault/unlock' to obtain the vault encryption key used for private recovery data decryption.
+     *
+     * This method calls PowerAuth Standard RESTful API endpoint `/pa/vault/unlock` to obtain the vault encryption key used for private recovery data decryption.
      *
      * @param authentication Authentication used for vault unlocking call.
      */
     activationRecoveryData(authentication: PowerAuthAuthentication): Promise<PowerAuthRecoveryActivationData>;
     /**
-     * Confirm given recovery code on the server.
+     * Confirm given recovery code on the server by calling a PowerAuth Standard RESTful API endpoint `/pa/recovery/confirm`.
+     *
      * The method is useful for situations when user receives a recovery information via OOB channel (for example via postcard).
      * Such recovery codes cannot be used without a proper confirmation on the server. To confirm codes, user has to authenticate himself
      * with a knowledge factor.
