+++ lib/PowerAuthTokenStore.js
@@ -49,7 +49,7 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
         if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
     }
 };
-import { __NativeWrapper } from "../internal/NativeWrapper";
+import { __NativeWrapper } from "./internal/NativeWrapper";
 /**
  * The PowerAuthTokenStore provides interface for managing access tokens. The class is using Keychain as
  * underlying storage for received data. Note that the whole store's interface is thread safe, but it's
@@ -93,7 +93,8 @@ var PowerAuthTokenStore = /** @class */ (function () {
         return this.wrapper.call("removeAllLocalTokens");
     };
     /**
-     * Create a new access token with given name for requested signature factors.
+     * Create a new access token with given name for requested signature factors by calling
+     * a PowerAuth Standard RESTful API endpoint `/pa/token/create`.
      *
      * Note that the method is thread safe, but it's not recommended to request for the same token
      * name in parallel when the token is not created yet. You can use hasLocalToken() method
@@ -118,7 +119,8 @@ var PowerAuthTokenStore = /** @class */ (function () {
         });
     };
     /**
-     * Remove previously created access token from the server and from local database.
+     * Remove previously created access token from the server and from local database by calling
+     * a PowerAuth Standard RESTful API endpoint `/pa/token/remove`.
      *
      * Note that if the removal request doesn't succeed, then the local token's data is not removed.
      * The method is thread safe, but it's not recommended to issue conflicting request for the same
